server/requirements.txt
python-dotenv~=1.0.0
fastapi~=0.109.0
pydantic~=2.5.3
motor~=3.3.2
pymongo~=4.6.1
uvicorn~=0.25.0
boto3~=1.34.18
botocore~=1.34.18
starlette~=0.35.1
python-multipart
python-telegram-bot
pika
